pyproject.toml
[project]
name = "project-cortex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.18",
    "flask>=3.1.2",
    "google-genai>=1.56.0",
    "openai>=2.14.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "tavily-python>=0.7.17",
    "trafilatura>=2.0.0",
    "youtube-transcript-api>=1.2.3",
]
